Extracted view logic to view controller
